==> jupyter_notebooks/03 - Data Cleaning.ipynb <==
# **Data Cleaning**

## Objectives

* Evaluate missing data
* Clean data

## Inputs

* outputs/datasets/collection/mushrooms.csv

## Outputs

* Generate cleaned Train and Test sets, saved under outputs/datasets/cleaned
* Data Cleaning Pipeline

## Additional Comments

* Will use `TargetEncoder` prior to eda so that Spearman and Pearson correlation coefficients can be obtained from categorical data


---

# Change working directory

* Need to change working directory from the current jupyter_notebooks folder to the parent folder in order to access the whole project

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

---

# Load collected data

import pandas as pd
df_raw_path = "outputs/datasets/collection/mushrooms.csv"
df = pd.read_csv(df_raw_path)
df.head()

Print out each of the columns to familiarize with their names.

for col in df.columns:
    print(col)

---

# Data Exploration

Check which columns have missing data - there is no columns missing data. 

vars_with_missing_data = df.columns[df.isna().sum() > 0].to_list()
vars_with_missing_data

from ydata_profiling import ProfileReport
if vars_with_missing_data:
    profile = ProfileReport(df=df[vars_with_missing_data], minimal=True)
    profile.to_notebook_iframe()
else:
    print("There are no variables with missing data")

# Correlation and PPS analysis

import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import ppscore as pps


def heatmap_corr(df, threshold, figsize=(20, 12), font_annot=8):
    """
    Plots a heatmap of variable correlation coefficients above the input threshold
    Taken from Walkthrough Project 02 - Churnometer

    Args:
        df: the DataFrame containing the dataset
        threshold: the minimum correlation coefficient to display in the heatmap
        figsize: the dimensions of the output figure 
        font_annot: the font size of the plots annotations

    Returns:
        None
    """
    if len(df.columns) > 1:
        mask = np.zeros_like(df, dtype=np.bool)
        mask[np.triu_indices_from(mask)] = True
        mask[abs(df) < threshold] = True

        fig, axes = plt.subplots(figsize=figsize)
        sns.heatmap(df, annot=True, xticklabels=True, yticklabels=True,
                    mask=mask, cmap='viridis', annot_kws={"size": font_annot}, ax=axes,
                    linewidth=0.5
                    )
        axes.set_yticklabels(df.columns, rotation=0)
        plt.ylim(len(df.columns), 0)
        plt.show()


def heatmap_pps(df, threshold, figsize=(20, 12), font_annot=8):
    """
    Plots a heatmap of PPS values above the input threshold
    Taken from Walkthrough Project 02 - Churnometer

    Args:
        df: the DataFrame containing the dataset
        threshold: the minimum pps value to display in the heatmap
        figsize: the dimensions of the output figure 
        font_annot: the font size of the plots annotations

    Returns:
        None
    """
    if len(df.columns) > 1:
        mask = np.zeros_like(df, dtype=np.bool)
        mask[abs(df) < threshold] = True
        fig, ax = plt.subplots(figsize=figsize)
        ax = sns.heatmap(df, annot=True, xticklabels=True, yticklabels=True,
                         mask=mask, cmap='rocket_r', annot_kws={"size": font_annot},
                         linewidth=0.05, linecolor='grey')
        plt.ylim(len(df.columns), 0)
        plt.show()


def CalculateCorrAndPPS(df, df_te):
    """
    Calculates correlation coefficients and PPS values between each pair of variables in the dataset
    Taken from Walkthrough Project 02 - Churnometer, altered to handle two different DataFrames

    Args:
        df: the DataFrame containing the dataset without encoding
        df_te: the DataFrame containing the dataset with target encoding

    Returns:
        df_corr_pearson: a DataFrame containing the correlation coefficients of each variable to 
                        `edible`, calculated by Pearson method 
        df_corr_spearman: a DataFrame containing the correlation coefficients of each variable to 
                        `edible`, calculated by Spearman method
        pps_matrix: a matrix of the power predictive score of each pair of variables in the dataset
    """
    df_corr_spearman = df_te.corr(method="spearman")
    df_corr_pearson = df_te.corr(method="pearson")

    pps_matrix_raw = pps.matrix(df)
    pps_matrix = pps_matrix_raw.filter(['x', 'y', 'ppscore']).pivot(columns='x', index='y', values='ppscore')

    pps_score_stats = pps_matrix_raw.query("ppscore < 1").filter(['ppscore']).describe().T
    print("PPS threshold - check PPS score IQR to decide threshold for heatmap \n")
    print(pps_score_stats.round(3))

    return df_corr_pearson, df_corr_spearman, pps_matrix


def DisplayCorrAndPPS(df_corr_pearson, df_corr_spearman, pps_matrix, CorrThreshold, PPS_Threshold,
                      figsize=(20, 12), font_annot=8):
    """
    Displays an output of all of the correlation coefficient and PPS plots
    Taken from Walkthrough Project 02 - Churnometer

    Args:
        df_corr_pearson: a DataFrame containing the Pearson calculated correlation coefficients for each variable in the dataset
        df_corr_spearman: a DataFrame containing the Spearman calculated correlation coefficients for each variable in the dataset
        pps_matrix: a matrix of the power predictive score of each pair of variables in the dataset
        CorrThreshold: the minimum correlation coefficients to display in the correlation coefficient heatmaps
        PPS_Threshold: the minimum power predictive score to display in the pps heatmap
        figsize: the dimensions of the heatmaps to produce
        font_annot: the font size of the annotations on the heatmaps

    Returns:
        None
    """
    print("\n")
    print("* Analyse how the target variable for your ML models are correlated with other variables (features and target)")
    print("* Analyse multi-colinearity, that is, how the features are correlated among themselves")

    print("\n")
    print("*** Heatmap: Spearman Correlation ***")
    print("It evaluates monotonic relationship \n")
    heatmap_corr(df=df_corr_spearman, threshold=CorrThreshold, figsize=figsize, font_annot=font_annot)

    print("\n")
    print("*** Heatmap: Pearson Correlation ***")
    print("It evaluates the linear relationship between two continuous variables \n")
    heatmap_corr(df=df_corr_pearson, threshold=CorrThreshold, figsize=figsize, font_annot=font_annot)

    print("\n")
    print("*** Heatmap: Power Predictive Score (PPS) ***")
    print(f"PPS detects linear or non-linear relationships between two columns.\n"
          f"The score ranges from 0 (no predictive power) to 1 (perfect predictive power) \n")
    heatmap_pps(df=pps_matrix, threshold=PPS_Threshold, figsize=figsize, font_annot=font_annot)

### Calculate Correlations and Power Predictive Score

We will encode the dataset with target encoding to calculate Pearson and Spearman correlations

from category_encoders import TargetEncoder
encoder = TargetEncoder()
df_te = df.copy()
cols = df.columns[df.dtypes=='object'].to_list()

for col in cols:
    df_te[col] = encoder.fit_transform(df[col], df['edible'])

df_te.head()

df_corr_pearson, df_corr_spearman, pps_matrix = CalculateCorrAndPPS(df, df_te)

Display on heatmaps

%matplotlib inline
DisplayCorrAndPPS(df_corr_pearson = df_corr_pearson,
                  df_corr_spearman = df_corr_spearman, 
                  pps_matrix = pps_matrix,
                  CorrThreshold = 0.5, PPS_Threshold =0.3,
                  figsize=(12,10), font_annot=10)

---

# Data Cleaning

As found previously, there is no explicitly missing or incorrectly formatted data. Therefore there is no need to perform any imputation on the dataset and it can be split into Train and Test sets as is.

### Split Train and Test Set

from sklearn.model_selection import train_test_split
TrainSet, TestSet, _, __ = train_test_split(df, df['edible'], test_size=0.2, random_state=0)

print(f"TrainSet shape: {TrainSet.shape} \nTestSet shape: {TestSet.shape}")

---

# Push Cleaned Data to Repo

import os
try:
    os.makedirs(name='outputs/datasets/cleaned')
except Exception as e:
    print(e)

### Train Set

TrainSet.to_csv("outputs/datasets/cleaned/TrainSetCleaned.csv", index=False)

### Test Set

TestSet.to_csv("outputs/datasets/cleaned/TestSetCleaned.csv", index=False)
